app/requirements.txt
psycopg2-binary==2.9.9
Django==4.2.13
djangorestframework==3.15.2
pytest==8.3.3
pytest-django==4.9.0